use size_t instead of int for array size in element bits

+++ src/t8_schemes/t8_default/t8_default_hex/t8_dhex_bits.c
@@ -26,7 +26,7 @@
 void
 t8_dhex_compute_reference_coords (const t8_dhex_t * elem,
                                   const double *ref_coords,
-                                  const int num_coords, double *out_coords)
+                                  const size_t num_coords, double *out_coords)
 {
   const p8est_quadrant_t *q1 = (const p8est_quadrant_t *) elem;
